backend/Module 1/register/requirements.txt
firebase-admin==6.2.0
pyrebase4==4.7.1
requests==2.26.0
flask==2.3.2
Flask-Cors==4.0.0